+++ src/frontend/js/components/AddressesManagement/AddressForm.tsx
@@ -1,14 +1,13 @@
 import { yupResolver } from '@hookform/resolvers/yup';
-import countries from 'i18n-iso-countries';
 import { Fragment, useEffect } from 'react';
 import { useForm } from 'react-hook-form';
 import { FormattedMessage, useIntl } from 'react-intl';
 import { messages } from 'components/AddressesManagement/index';
-import { CheckboxField, SelectField, TextField } from 'components/Form';
+import { CheckboxField, TextField } from 'components/Form';
+import { CountrySelectField } from 'components/Form/CountrySelectField';
 import { useAddresses } from 'hooks/useAddresses';
 import type { Address } from 'types/Joanie';
-import { Maybe } from 'types/utils';
-import validationSchema, { ErrorKeys, errorMessages } from './validationSchema';
+import validationSchema, { getLocalizedErrorMessage } from './validationSchema';
 
 export type AddressFormValues = Omit<Address, 'id' | 'is_main'> & { save: boolean };
 
@@ -39,27 +38,6 @@ const AddressForm = ({ handleReset, onSubmit, address }: Props) => {
 
   const addresses = useAddresses();
   const intl = useIntl();
-  const [languageCode] = intl.locale.split('-');
-  const countryList = countries.getNames(languageCode);
-
-  const getLocalizedErrorMessage = (
-    error: Maybe<
-      | string
-      | {
-          key: ErrorKeys;
-          values: Record<PropertyKey, string | number | Array<string | number>>;
-        }
-    >,
-  ) => {
-    if (!error) return undefined;
-
-    if (typeof error === 'string' || errorMessages[error.key] === undefined) {
-      // If the error has not been translated we return a default error message.
-      return intl.formatMessage(errorMessages[ErrorKeys.MIXED_INVALID]);
-    }
-
-    return intl.formatMessage(errorMessages[error.key], error.values);
-  };
 
   /**
    * Prevent form to be submitted and clear `editedAddress` state.
@@ -85,7 +63,7 @@ const AddressForm = ({ handleReset, onSubmit, address }: Props) => {
         id="title"
         label={intl.formatMessage(messages.titleInputLabel)}
         error={!!formState.errors.title}
-        message={getLocalizedErrorMessage(formState.errors.title?.message)}
+        message={getLocalizedErrorMessage(intl, formState.errors.title?.message)}
         {...register('title')}
       />
       <div className="form-group">
@@ -95,7 +73,7 @@ const AddressForm = ({ handleReset, onSubmit, address }: Props) => {
           id="first_name"
           label={intl.formatMessage(messages.first_nameInputLabel)}
           error={!!formState.errors.first_name}
-          message={getLocalizedErrorMessage(formState.errors.first_name?.message)}
+          message={getLocalizedErrorMessage(intl, formState.errors.first_name?.message)}
           {...register('first_name')}
         />
         <TextField
@@ -104,7 +82,7 @@ const AddressForm = ({ handleReset, onSubmit, address }: Props) => {
           id="last_name"
           label={intl.formatMessage(messages.last_nameInputLabel)}
           error={!!formState.errors.last_name}
-          message={getLocalizedErrorMessage(formState.errors.last_name?.message)}
+          message={getLocalizedErrorMessage(intl, formState.errors.last_name?.message)}
           {...register('last_name')}
         />
       </div>
@@ -114,7 +92,7 @@ const AddressForm = ({ handleReset, onSubmit, address }: Props) => {
         id="address"
         label={intl.formatMessage(messages.addressInputLabel)}
         error={!!formState.errors.address}
-        message={getLocalizedErrorMessage(formState.errors.address?.message)}
+        message={getLocalizedErrorMessage(intl, formState.errors.address?.message)}
         {...register('address')}
       />
       <div className="form-group">
@@ -124,7 +102,7 @@ const AddressForm = ({ handleReset, onSubmit, address }: Props) => {
           id="postcode"
           label={intl.formatMessage(messages.postcodeInputLabel)}
           error={!!formState.errors.postcode}
-          message={getLocalizedErrorMessage(formState.errors.postcode?.message)}
+          message={getLocalizedErrorMessage(intl, formState.errors.postcode?.message)}
           {...register('postcode')}
         />
         <TextField
@@ -133,28 +111,19 @@ const AddressForm = ({ handleReset, onSubmit, address }: Props) => {
           id="city"
           label={intl.formatMessage(messages.cityInputLabel)}
           error={!!formState.errors.city}
-          message={getLocalizedErrorMessage(formState.errors.city?.message)}
+          message={getLocalizedErrorMessage(intl, formState.errors.city?.message)}
           {...register('city')}
         />
       </div>
-      <SelectField
+      <CountrySelectField
         aria-invalid={!!formState.errors.country}
         aria-required={false}
         id="country"
         label={intl.formatMessage(messages.countryInputLabel)}
         error={!!formState.errors.country}
-        message={getLocalizedErrorMessage(formState.errors.country?.message)}
+        message={getLocalizedErrorMessage(intl, formState.errors.country?.message)}
         {...register('country', { value: address?.country, required: true })}
-      >
-        <option disabled value="-">
-          -
-        </option>
-        {Object.entries(countryList).map(([value, label]) => (
-          <option key={`address-countryList-${value}`} value={value}>
-            {label}
-          </option>
-        ))}
-      </SelectField>
+      />
       {!address ? (
         <CheckboxField
           aria-invalid={!!formState.errors?.save}
@@ -162,7 +131,7 @@ const AddressForm = ({ handleReset, onSubmit, address }: Props) => {
           id="save"
           label={intl.formatMessage(messages.saveInputLabel)}
           error={!!formState.errors?.save}
-          message={getLocalizedErrorMessage(formState.errors.save?.message)}
+          message={getLocalizedErrorMessage(intl, formState.errors.save?.message)}
           {...register('save')}
         />
       ) : null}