+++ CHANGELOG.md
@@ -33,6 +33,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   to always display value with a dot no matter active language
 - Language choices should be ordered alphabetically in course run admin form
 
+### Added
+
+- Dashboard addresses management components 
+
 ## [2.16.0]
 
 ### Fixed